requirements.txt
opencv-python       4.10.0.84
pyzmq               26.2.0
python-dotenv       1.0.1
requests            2.32.3